requirements.txt
# Core
torch>=2.2
torchvision>=0.17
numpy>=1.24
Pillow>=10.0
tqdm>=4.66
PyYAML>=6.0
rich>=13.7
opencv-python>=4.8

# Optional (for FLOPs/latency & calibration plots)
thop>=0.1.1.post2209072238
matplotlib>=3.8
